Refactors makefile to use more explicit build and install targets (#1563)
- Updates contributing docs to reflect this
- Adds TCE_RELEASE_DIR to makefile which is where tanzu framework is built

+++ hack/build-tanzu.sh
@@ -14,8 +14,13 @@ TANZU_FRAMEWORK_REPO_HASH="${TANZU_FRAMEWORK_REPO_HASH:-""}"
 # all of tanzu-framework's environments
 ENVS="${ENVS:-""}"
 
+if [[ -z "${TCE_RELEASE_DIR}" ]]; then
+    echo "TCE_RELEASE_DIR is not set"
+    exit 1
+fi
+
 # Change directories to a clean build space
-ROOT_REPO_DIR="/tmp/tce-release"
+ROOT_REPO_DIR="${TCE_RELEASE_DIR}"
 rm -fr "${ROOT_REPO_DIR}"
 mkdir -p "${ROOT_REPO_DIR}"
 cd "${ROOT_REPO_DIR}" || exit 1
@@ -53,17 +58,9 @@ make set-unstable-versions
 fi
 # generate the correct tkg-bom (which references the tkr-bom)
 # make configure-bom
-# build and install all "tanzu-framework" CLI plugins
+# build all "tanzu-framework" CLI plugins
 # (e.g. management-cluster, cluster, etc)
 BUILD_SHA=${BUILD_SHA} BUILD_VERSION=${FRAMEWORK_BUILD_VERSION} make ENVS="${ENVS}" clean-catalog-cache clean-cli-plugins build-cli
-
-# Only do an install if the environments to build contain the current host OS.
-# The tanzu-framework `build-install-cli-all` target always uses the current host OS, and if that's not being built it will fail.
-GOHOSTOS=$(go env GOHOSTOS)
-if [[ "$ENVS" == *"${GOHOSTOS}"* ]]; then
-BUILD_SHA=${BUILD_SHA} BUILD_VERSION=${FRAMEWORK_BUILD_VERSION} make ENVS="${ENVS}" install-cli-plugins install-cli
-fi
-
 
 # by default, tanzu-framework only builds admins plugins for the current platform. we need darwin also.
 BUILD_VERSION=${FRAMEWORK_BUILD_VERSION} GOHOSTOS=linux GOHOSTARCH=amd64 make ENVS="${ENVS}" build-plugin-admin
